requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
python-dotenv==0.15.0
requests==2.25.1
paho-mqtt==1.6.1